chore: Refactor interactive mode to a module
